Rename `render` package to `report`.
