+++ composite.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/dogmatiq/testkit/assert"
 	"github.com/dogmatiq/testkit/engine/fact"
-	"github.com/dogmatiq/testkit/render"
+	"github.com/dogmatiq/testkit/report"
 )
 
 // AllOf is an expectation that passes only if all of its children pass.
@@ -191,7 +191,7 @@ func (e *compositeExpectation) Ok() bool {
 // ok is true if the expectation is considered to have passed. This may not be
 // the same value as returned from Ok() when this expectation is used as a child
 // of a composite expectation.
-func (e *compositeExpectation) BuildReport(ok bool, r render.Renderer) *assert.Report {
+func (e *compositeExpectation) BuildReport(ok bool, r report.Renderer) *assert.Report {
 	e.Ok() // populate e.ok and e.outcome
 
 	rep := &assert.Report{
