--- render/doc.go
@@ -1,3 +1,0 @@
-// Package render contains utilities for rendering various values in a
-// human-readable manner.
-package render
