+++ report/doc.go
@@ -1,0 +1,2 @@
+// Package report builds human-readable reports that explain test results.
+package report
