+++ report/renderer_test.go
@@ -1,4 +1,4 @@
-package render
+package report
 
 import (
 	"io"
